+++ vsc-extension/src-views/components/TestCaseEditor.tsx
@@ -1,12 +1,11 @@
-import { VSCodeButton } from "@vscode/webview-ui-toolkit/react";
-import React, { FC, useCallback } from "react";
+import React, { FC } from "react";
 import {
   FunctionTestCase,
   SelectedFunction,
   SerializableFunctionDeclaration,
 } from "../../src-shared/source-info";
-import { useExtensionState } from "./ExtensionState";
 import { ParameterValueEditor } from "./ParameterValueEditor";
+import { RunButton } from "./RunButton";
 
 interface Props {
   selectedFunctionInfo: SerializableFunctionDeclaration | undefined;
@@ -51,29 +50,4 @@ const TemperatureEditor: FC = () => {
       <p>(temperature editor here)</p>
     </div>
   );
-};
-
-const RunButton: FC<{
-  selectedFunction: SelectedFunction;
-  testCaseIndex: number;
-}> = ({ selectedFunction, testCaseIndex }) => {
-  const { rpcProvider } = useExtensionState();
-  const { fileName, functionName } = selectedFunction;
-
-  const onRun = useCallback(async () => {
-    try {
-      await rpcProvider?.rpc("runTestCase", {
-        fileName,
-        functionName,
-        testCaseIndex,
-      });
-    } catch (ex) {
-      console.error(`Failure to run: ${ex}`, ex);
-    }
-  }, [fileName, functionName, rpcProvider, testCaseIndex]);
-
-  if (!rpcProvider) {
-    return null;
-  }
-  return <VSCodeButton onClick={onRun}>Run</VSCodeButton>;
 };
